Documentation complete + minor refactors.
