+++ server/games/i-agent.js
@@ -1,8 +1,6 @@
 class IAgent {
     /** @type {String} The token of the current agent. */
     token;
-    /** @type {{ event: String, args: []}[]} Agent event record */
-    events = [];
     /** @type {Function} Action resolver */
     pending;
