+++ server/games/love-letter/agent.js
@@ -24,6 +24,9 @@ class Agent extends IAgent {
         return Action[action.action](...action.params);
     }
 
+    /**
+     * @returns The state corresponding to this agent.
+     */
     getState() {
         const agentState = {...this.state};
         delete agentState.agents;
